% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasi.R
\name{calculate_hasi_r}
\alias{calculate_hasi_r}
\title{Calculates the HASI-R score for each patient and visit.}
\usage{
calculate_hasi_r(
  patientID,
  visitDY,
  BodySite,
  BSA,
  InflammColorChg,
  Induration,
  OpenSkinSurface,
  Tunnels
)
}
\arguments{
\item{patientID}{Character vector representing patient IDs.}

\item{visitDY}{Integer vector representing visit days.}

\item{BodySite}{Character vector representing body sites.}

\item{BSA}{Integer vector representing body surface area scores (0-6).}

\item{InflammColorChg}{Integer vector representing inflammatory color change scores (0-3).}

\item{Induration}{Integer vector representing induration scores (0-3).}

\item{OpenSkinSurface}{Integer vector representing open skin surface scores (0-3).}

\item{Tunnels}{Integer vector representing tunnels scores (0-3).}
}
\value{
A data frame with patientID, visitDY, and the calculated HASI-R score.
}
\description{
Calculates the HASI-R score for each patient and visit.
}
\examples{

patientID <-
  c(
    "000-001", "000-001", "000-001", "000-001",
    "000-002", "000-002", "000-002", "000-002"
  )
visitDY <- c(1, 1, 2, 2, 1, 1, 2, 2)
BodySite <-
  c(
    "Axillae", "Groin", "Axillae", "Groin",
    "Axillae", "Groin", "Axillae", "Groin"
  )
BSA <- c(1, 2, 1, 2, 2, 1, 2, 1)
InflammColorChg <- c(1, 2, 1, 2, 2, 1, 2, 1)
Induration <- c(2, 2, 1, 1, 1, 1, 1, 1)
OpenSkinSurface <- c(1, 1, 2, 3, 2, 3, 2, 3)
Tunnels <- c(0, 1, 2, 0, 2, 0, 2, 0)

# hasi_r_scores <-
#   calculate_hasi_r(
#     patientID, visitDY, BodySite,
#     InflammColorChg, Induration, OpenSkinSurface,
#     Tunnels
#   )
# hasi_r_scores

}
\references{
Goldfarb N, Lowes MA, Butt M, King T, Alavi A, Kirby JS.
Hidradenitis Suppurativa Area and Severity Index Revised (HASI-R):
psychometric property assessment. Br J Dermatol. 2021 May;184(5):905-912.
doi: 10.1111/bjd.19565. Epub 2020 Dec 30. PMID: 32969027; PMCID:
PMC8573730.
}
