% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa_ordinal_scale.R
\name{convert_bsa_to_ordinal}
\alias{convert_bsa_to_ordinal}
\title{Converts BSA percentage to the ordinal scale for HASI-R scoring.}
\usage{
convert_bsa_to_ordinal(bsa_percent)
}
\arguments{
\item{bsa_percent}{Numeric vector representing BSA percentages.}
}
\value{
An integer vector representing the ordinal scale values.
}
\description{
Converts BSA percentage to the ordinal scale for HASI-R scoring.
}
\examples{

bsa_percent <- c(0, 2, 5, 12, 25, 40, 75)
bsa_ordinal <- convert_bsa_to_ordinal(bsa_percent)
bsa_ordinal

}
