% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasi_bsa_to_ordinal.R
\name{hasi_bsa_to_ordinal}
\alias{hasi_bsa_to_ordinal}
\title{Converts BSA percentage to the ordinal scale for HASI-R scoring.}
\usage{
hasi_bsa_to_ordinal(
  bsa_percent_total_body = NULL,
  bsa_percent_within_site = NULL,
  bsa_ordinal = NULL,
  bodysite = NULL
)
}
\arguments{
\item{bsa_percent_total_body}{Numeric vector representing BSA percentages. (0-15, depending on site)}

\item{bsa_percent_within_site}{Numeric vector representing BSA percentages. (0-100)}

\item{bsa_ordinal}{Numeric vector representing BSA values. (0-6)}

\item{bodysite}{Optional character vector representing body sites.}
}
\value{
An integer vector representing the ordinal scale values.
}
\description{
Converts BSA percentage to the ordinal scale for HASI-R scoring.
}
\examples{

bsa_percent <- c(0, 2, 5, 12, 25, 40, 75)
bsa_ordinal <- hasi_bsa_to_ordinal(bsa_percent)
bsa_ordinal

}
