% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasi.R
\name{hasi_r_num}
\alias{hasi_r_num}
\title{Calculates the HASI-R score for each patient and visit.}
\usage{
hasi_r_num(
  bsa_percent_within_site = NULL,
  bsa_percent_total_body = NULL,
  bsa_ordinal = NULL,
  bodysite = NULL,
  inflam_color_chg,
  induration,
  open_skin_surface,
  tunnels
)
}
\arguments{
\item{bsa_percent_within_site}{Numeric vector representing BSA percentages. (0-100)}

\item{bsa_percent_total_body}{Numeric vector representing BSA percentages. (0-15, depending on site)}

\item{bsa_ordinal}{Numeric vector representing BSA values. (0-6)}

\item{bodysite}{Optional character vector representing body sites.}

\item{inflam_color_chg}{Integer vector representing inflammatory color change scores (0-3).}

\item{induration}{Integer vector representing induration scores (0-3).}

\item{open_skin_surface}{Integer vector representing open skin surface scores (0-3).}

\item{tunnels}{Integer vector representing tunnels scores (0-3).}
}
\value{
A data frame with patientID, visitDY, and the calculated HASI-R score.
}
\description{
Calculates the HASI-R score for each patient and visit.
}
\examples{

patientID <-
  c(
    "000-001", "000-001", "000-001", "000-001",
    "000-002", "000-002", "000-002", "000-002"
  )
visitDY <- c(1, 1, 2, 2, 1, 1, 2, 2)
BodySite <-
  c(
    "Left Axilla", "Chest", "Left Axilla", "Chest",
    "Left Axilla", "Chest", "Left Axilla", "Chest"
  )
BSA <- c(1, 2, 1, 2, 2, 1, 2, 1)
InflammColorChg <- c(1, 2, 1, 2, 2, 1, 2, 1)
Induration <- c(2, 2, 1, 1, 1, 1, 1, 1)
OpenSkinSurface <- c(1, 1, 2, 3, 2, 3, 2, 3)
Tunnels <- c(0, 1, 2, 0, 2, 0, 2, 0)

# hasi_r_scores <-
#   hasi_r_num(
#     patientID, visitDY, BodySite, BSA,
#     InflammColorChg, Induration, OpenSkinSurface,
#     Tunnels
#   )
# hasi_r_scores

}
\references{
Goldfarb N, Lowes MA, Butt M, King T, Alavi A, Kirby JS.
Hidradenitis Suppurativa Area and Severity Index Revised (HASI-R):
psychometric property assessment. Br J Dermatol. 2021 May;184(5):905-912.
doi: 10.1111/bjd.19565. Epub 2020 Dec 30. PMID: 32969027; PMCID:
PMC8573730.
}
