% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasi.R
\name{hs_hasi_r}
\alias{hs_hasi_r}
\title{Calculates the HASI-R score for Hidradenitis Suppurativa severity.}
\usage{
hs_hasi_r(nodules, abscesses, fistulas)
}
\arguments{
\item{nodules}{List of integer vectors representing the count of inflammatory nodules in each body region.}

\item{abscesses}{List of integer vectors representing the count of abscesses in each body region.}

\item{fistulas}{List of integer vectors representing the count of draining fistulas in each body region.}

\item{region}{?}
}
\value{
A numeric value representing the HASI-R score.
}
\description{
Calculates the HASI-R score for Hidradenitis Suppurativa severity.
}
\examples{

nodules <- list(axillae = 3, groin = 2, buttocks = 0)
abscesses <- list(axillae = 1, groin = 1, buttocks = 0)
fistulas <- list(axillae = 1, groin = 0, buttocks = 1)
hasi_r_score <- hs_hasi_r(nodules, abscesses, fistulas)

}
