% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHS4_char.R
\name{IHS4_char}
\alias{IHS4_char}
\title{Converts IHS4 numerical scores to character values.}
\usage{
IHS4_char(IHS4_scores)
}
\arguments{
\item{hs_pga_scores}{An integer vector representing the IHS4 scores (0-3: Mild, 4-10: Moderate, 11- : Severe).}
}
\value{
A character vector representing the corresponding IHS4 categories.
}
\description{
Converts IHS4 numerical scores to character values.
}
\examples{
IHS4_scores <- c(2, 5, 12)

IHS4_char(IHS4_scores)

}
