% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHS4_char.R
\name{IHS4_char}
\alias{IHS4_char}
\title{Converts IHS4 numerical scores to character values.}
\usage{
IHS4_char(IHS4_scores)
}
\arguments{
\item{IHS4_scores}{An integer vector representing the IHS4 scores (0-3: Mild, 4-10: Moderate, >=11: Severe).}
}
\value{
A character vector representing the corresponding IHS4 categories.
}
\description{
Converts IHS4 numerical scores to character values.
}
\examples{
IHS4_char(c(2, 5, 12))
}
\references{
Zouboulis, C C et al. “Development and validation of the
International Hidradenitis Suppurativa Severity Score System (IHS4),
a novel dynamic scoring system to assess HS severity.” The British
journal of dermatology vol. 177,5 (2017): 1401-1409. doi:10.1111/bjd.15748
}
